requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.3.1
colorama==0.4.6
fastapi==0.123.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.1
pi==0.1.2
playwright==1.49.0
pluggy==1.5.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.0.0
PySocks==1.7.1
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.6.2
python-dotenv==1.0.1
python-slugify==8.0.4
python-telegram-bot==21.6
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.50.0
text-unidecode==1.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.38.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
